+++ src/main/kotlin/app/hinl/oreharvest/screenhandler/ButtonEntry.kt
@@ -2,4 +2,9 @@ package app.hinl.oreharvest.screenhandler
 
 import net.minecraft.text.Text
 
-data class ButtonEntry(val isChecked: Boolean, val buttonText: Text, val buttonAction: (Boolean) -> Unit)
+data class ButtonEntry(
+    val isChecked: Boolean,
+    val buttonText: Text? = null,
+    val buttonToolTips: Text? = null,
+    val buttonAction: (Boolean) -> Unit
+)