+++ src/main/kotlin/app/hinl/oreharvest/screenhandler/ChestScreen.kt
@@ -1,10 +1,13 @@
 package app.hinl.oreharvest.screenhandler
 
 import app.hinl.oreharvest.Constants
+import com.mojang.blaze3d.platform.GlStateManager.DstFactor
+import com.mojang.blaze3d.platform.GlStateManager.SrcFactor
 import com.mojang.blaze3d.systems.RenderSystem
+import net.minecraft.client.MinecraftClient
 import net.minecraft.client.gui.screen.ingame.HandledScreen
 import net.minecraft.client.gui.screen.ingame.ScreenHandlerProvider
-import net.minecraft.client.gui.widget.CheckboxWidget
+import net.minecraft.client.gui.widget.ButtonWidget
 import net.minecraft.client.render.GameRenderer
 import net.minecraft.client.util.math.MatrixStack
 import net.minecraft.entity.player.PlayerInventory
@@ -17,7 +20,7 @@ class ChestScreen(handler: ChestScreenHandler, inventory: PlayerInventory?, titl
     private val texture = Identifier(Constants.ModID, "textures/screen/chest_base.png")
     private val rows: Int
     private val buttonEntries = handler.buttonEntries
-    private val buttonList = mutableListOf<CheckboxWidget>()
+    private val buttonList = mutableListOf<OptionCheckbox>()
 
     init {
         rows = handler.rows
@@ -38,7 +41,11 @@ class ChestScreen(handler: ChestScreenHandler, inventory: PlayerInventory?, titl
                 width = 12,
                 height = 12,
                 message = buttonEntry.buttonText,
-                checked = buttonEntry.isChecked,
+                tooltipSupplier = { _: ButtonWidget, matrixStack: MatrixStack, mouseX: Int, mouseY: Int ->
+                    renderTooltip(matrixStack, buttonEntry.buttonToolTips, mouseX, mouseY)
+                },
+                isShowMessage = true,
+                isChecked = buttonEntry.isChecked,
                 clickListener = buttonEntry.buttonAction
             )
             buttonList.add(button)
@@ -72,18 +79,62 @@ class ChestScreen(handler: ChestScreenHandler, inventory: PlayerInventory?, titl
     }
 
     class OptionCheckbox(
+        private var isChecked: Boolean,
+        private val isShowMessage: Boolean = false,
         x: Int,
         y: Int,
         width: Int,
         height: Int,
         message: Text?,
-        checked: Boolean,
-        showMessage: Boolean = false,
+        tooltipSupplier: TooltipSupplier = EMPTY,
         private val clickListener: (Boolean) -> Unit
-    ) : CheckboxWidget(x, y, width, height, message, checked, showMessage) {
-        override fun onClick(mouseX: Double, mouseY: Double) {
-            super.onClick(mouseX, mouseY)
+    ) : ButtonWidget(
+        x, y, width, height, message,
+        {
+        }, tooltipSupplier
+    ) {
+        private val texture = Identifier(Constants.ModID, "textures/screen/check_box.png")
+
+        override fun onPress() {
+            super.onPress()
+            isChecked = isChecked.not()
             clickListener.invoke(isChecked)
+        }
+
+        override fun renderButton(matrices: MatrixStack?, mouseX: Int, mouseY: Int, delta: Float) {
+            val minecraftClient = MinecraftClient.getInstance()
+            RenderSystem.setShaderTexture(0, texture)
+            RenderSystem.enableDepthTest()
+            val textRenderer = minecraftClient.textRenderer
+            RenderSystem.setShaderColor(1.0f, 1.0f, 1.0f, alpha)
+            RenderSystem.enableBlend()
+            RenderSystem.defaultBlendFunc()
+            RenderSystem.blendFunc(SrcFactor.SRC_ALPHA, DstFactor.ONE_MINUS_SRC_ALPHA)
+            drawTexture(
+                matrices,
+                x,
+                y,
+                if (this.isFocused) 12.0f else 0.0f,
+                if (isChecked) 12.0f else 0.0f,
+                width,
+                height,
+                64,
+                64
+            )
+            renderBackground(matrices, minecraftClient, mouseX, mouseY)
+            if (this.isHovered) {
+                renderTooltip(matrices, mouseX, mouseY)
+            }
+            if (isShowMessage && message != null) {
+                drawTextWithShadow(
+                    matrices,
+                    textRenderer,
+                    message,
+                    x + 14,
+                    y,
+                    14737632
+                )
+            }
         }
     }
 }