+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Disable button on the chest UI
 ### Changed
 ### Fixed
+- Growth cocoa will cause crash issue
 
 ## [1.0.3] - 2022-07-28
 ### Added